Understand the experimental fringes now
Fringes occur because the beam is no longer fully polarized after the first MWP even for a perfect beam in the simulations due to noise. IRL the fringes are much more pronounced because the depolarization is higher. Therefore we need a density matrix approach to the MWP action to get the A and B polarizations to understand the spin texture.